+++ app/src/main/java/campus/tech/kakao/map/di/RepositoryModule.kt
@@ -1,0 +1,20 @@
+package campus.tech.kakao.map.di
+
+import campus.tech.kakao.map.repository.Repository
+import campus.tech.kakao.map.repository.RepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+
+    @Binds
+    @Singleton
+    abstract fun bindRepository(
+        repositoryImpl: RepositoryImpl
+    ): Repository
+}
